Allow to define menu in deletedialog.

+++ pimcommon/storageservice/dialog/storageservicedeletedialog.cpp
@@ -27,6 +27,7 @@
 #include <KSharedConfig>
 #include <KMessageBox>
 #include <KFileDialog>
+#include <KMenu>
 
 #include <QGridLayout>
 #include <QLabel>
@@ -38,12 +39,54 @@
 
 using namespace PimCommon;
 
-StorageServiceDeleteDialog::StorageServiceDeleteDialog(PimCommon::StorageServiceAbstract *storage, QWidget *parent)
+StorageServiceDeleteTreeWidget::StorageServiceDeleteTreeWidget(PimCommon::StorageServiceDeleteDialog::DeleteType type, PimCommon::StorageServiceAbstract *storageService, QWidget *parent)
+    : PimCommon::StorageServiceTreeWidget(storageService, parent),
+      mDeleteType(type)
+{
+}
+
+StorageServiceDeleteDialog::DeleteType StorageServiceDeleteTreeWidget::deleteType() const
+{
+    return mDeleteType;
+}
+
+void StorageServiceDeleteTreeWidget::createMenuActions(KMenu *menu)
+{
+    StorageServiceTreeWidget::createMenuActions(menu);
+    const PimCommon::StorageServiceTreeWidget::ItemType type = StorageServiceTreeWidget::itemTypeSelected();
+
+    switch(mDeleteType) {
+    case PimCommon::StorageServiceDeleteDialog::DeleteAll:
+        if (type != StorageServiceTreeWidget::UnKnown)
+            menu->addAction(KIcon(QLatin1String("edit-delete")), i18n("Delete"), this, SIGNAL(deleteFileFolder()));
+        break;
+    case PimCommon::StorageServiceDeleteDialog::DeleteFiles:
+        if (type == StorageServiceTreeWidget::File)
+            menu->addAction(KIcon(QLatin1String("edit-delete")), i18n("Delete File"), this, SIGNAL(deleteFileFolder()));
+        break;
+    case PimCommon::StorageServiceDeleteDialog::DeleteFolders:
+        if (type != StorageServiceTreeWidget::Folder)
+            menu->addAction(KIcon(QLatin1String("edit-delete")), i18n("Delete Folder"), this, SIGNAL(deleteFileFolder()));
+        break;
+    }
+}
+
+StorageServiceDeleteDialog::StorageServiceDeleteDialog(DeleteType type, PimCommon::StorageServiceAbstract *storage, QWidget *parent)
     : KDialog(parent),
+      mDeleteType(type),
       mStorage(storage)
 {
-    setCaption( i18n( "Delete File" ) );
-
+    switch(mDeleteType) {
+    case DeleteAll:
+        setCaption(i18n("Delete"));
+        break;
+    case DeleteFiles:
+        setCaption(i18n("Delete Files"));
+        break;
+    case DeleteFolders:
+        setCaption(i18n("Delete Folders"));
+        break;
+    }
     mStorageServiceProgressIndicator = new PimCommon::StorageServiceProgressIndicator(this);
     connect(mStorageServiceProgressIndicator, SIGNAL(updatePixmap(QPixmap)), this, SLOT(slotUpdatePixmap(QPixmap)));
 
@@ -55,13 +98,13 @@ StorageServiceDeleteDialog::StorageServiceDeleteDialog(PimCommon::StorageService
 
     QHBoxLayout *hbox = new QHBoxLayout;
     vbox->addLayout(hbox);
-    QLabel *lab = new QLabel(i18n("Select file to delete:"));
+    QLabel *lab = new QLabel(i18n("Select file/folder to delete:"));
     hbox->addWidget(lab);
     mLabelProgressIncator = new QLabel;
     hbox->addWidget(mLabelProgressIncator);
     hbox->setAlignment(mLabelProgressIncator, Qt::AlignLeft);
-    mTreeWidget = new StorageServiceTreeWidget(storage);
-
+    mTreeWidget = new StorageServiceDeleteTreeWidget(mDeleteType, storage);
+    connect(mTreeWidget, SIGNAL(deleteFileFolder()), this, SLOT(slotDelete()));
 
     vbox->addWidget(mTreeWidget);
     connect(mTreeWidget, SIGNAL(itemDoubleClicked(QTreeWidgetItem*,int)), this, SLOT(slotItemDoubleClicked(QTreeWidgetItem*,int)));
@@ -175,9 +218,11 @@ void StorageServiceDeleteDialog::slotDelete()
     StorageServiceTreeWidgetItem *storageServiceItem = dynamic_cast<StorageServiceTreeWidgetItem*>(mTreeWidget->currentItem());
     if (storageServiceItem) {
         if (mTreeWidget->type(storageServiceItem) == StorageServiceTreeWidget::File) {
-            deleteFile(storageServiceItem);
+            if ((mDeleteType == DeleteFiles) || (mDeleteType == DeleteAll))
+                deleteFile(storageServiceItem);
         } else if (mTreeWidget->type(storageServiceItem) == StorageServiceTreeWidget::Folder) {
-            deleteFolder(storageServiceItem);
+            if ((mDeleteType == DeleteFolders) || (mDeleteType == DeleteAll))
+                deleteFolder(storageServiceItem);
         }
     }
 }
